pyproject.toml
[tool.poetry]
name = "имя_проекта"
version = "0.1.0"
description = "Краткое описание вашего проекта"
authors = ["Ваше Имя <ваш.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
uvicorn = "^0.30.6"
fastapi = "^0.115.0"
python-dotenv = "^1.0.1"
pydantic = "^2.9.2"
logging = "^0.4.9.6"
asyncpg = "^0.29.0"
aiohttp = "^3.10.5"
asyncio = "^3.4.3"
yoyo-migrations = "^9.0.0"
psycopg2-binary = "^2.9.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
